MIME4J-72: Removed code clutter but left the general disposal mechanism as it is Contributed by Markus Wiederkehr
git-svn-id: https://svn.apache.org/repos/asf/james/mime4j/trunk@703658 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/java/org/apache/james/mime4j/message/TempFileBinaryBody.java
@@ -66,9 +66,6 @@ class TempFileBinaryBody extends AbstractBody implements BinaryBody {
      * @see org.apache.james.mime4j.message.Body#writeTo(java.io.OutputStream, int)
      */
     public void writeTo(OutputStream out, int mode) throws IOException {
-        if (disposed)
-            throw new IllegalStateException("TempFileBinaryBody has been disposed");
-
         final InputStream inputStream = getInputStream();
         CodecUtil.copy(inputStream,out);
     }
@@ -79,12 +76,10 @@ class TempFileBinaryBody extends AbstractBody implements BinaryBody {
      * @see org.apache.james.mime4j.message.Disposable#dispose()
      */
     public void dispose() {
-        try {
+        if (tempFile != null) {
             tempFile.delete();
-        } finally {
             tempFile = null;
-
-            super.dispose();
         }
     }
+
 }
